RESTAPI 구축 완료

+++ src/main/java/site/metacoding/white/service/BoardService.java
@@ -11,9 +11,11 @@ import lombok.RequiredArgsConstructor;
 import site.metacoding.white.domain.Board;
 import site.metacoding.white.domain.BoardRepository;
 import site.metacoding.white.dto.BoardReqDto.BoardSaveReqDto;
+import site.metacoding.white.dto.BoardReqDto.BoardUpdateReqDto;
 import site.metacoding.white.dto.BoardRespDto.BoardAllRespDto;
 import site.metacoding.white.dto.BoardRespDto.BoardDetailRespDto;
 import site.metacoding.white.dto.BoardRespDto.BoardSaveRespDto;
+import site.metacoding.white.dto.BoardRespDto.BoardUpdateRespDto;
 
 // 트랜잭션 관리
 // DTO 변환해서 컨트롤러에게 돌려줘야함
@@ -49,12 +51,15 @@ public class BoardService {
 	}
 
 	@Transactional
-	public void update(Long id, Board board) {
+	public BoardUpdateRespDto update(BoardUpdateReqDto boardUpdateReqDto) { // 찾아 가져와서 영속성 컨텍스트에 집어 넣은 것
+		Long id = boardUpdateReqDto.getId();
 		Optional<Board> boardOP = boardRepository.findById(id);
 		if (boardOP.isPresent()) {
-			boardOP.get().update(board.getTitle(), board.getContent());
+			Board boardPS = boardOP.get();
+			boardPS.update(boardUpdateReqDto.getTitle(), boardUpdateReqDto.getContent());
+			return new BoardUpdateRespDto(boardPS);
 		} else {
-			throw new RuntimeException("해당 " + id + "로 수정 할 수 없습니다.");
+			throw new RuntimeException("해당 " + id + "로 수정을 할 수 없습니다.");
 		}
 
 	} // 트랜잭션 종료시 -> 더티체킹을 함
@@ -73,7 +78,13 @@ public class BoardService {
 
 	@Transactional
 	public void deleteById(Long id) {
-		boardRepository.deleteById(id);
+		Optional<Board> boardOP = boardRepository.findById(id);
+		if (boardOP.isPresent()) {
+			boardRepository.deleteById(id);
+		} else {
+			throw new RuntimeException("해당 " + id + "로 삭제를 할 수 없습니다.");
+		}
+
 	}
 
 }
